Complete user profile and webpack setting

+++ src/controllers/videoController.js
@@ -8,7 +8,7 @@ export const home = async (req, res) => {
 
 export const watch = async (req, res) => {
     const { id } = req.params;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("owner");
 
     if (!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." });
@@ -16,23 +16,35 @@ export const watch = async (req, res) => {
 
     const owner = await User.findById(video.owner);
 
-    return res.render("watch", { pageTitle: video.title, video, owner });
+    return res.render("watch", { pageTitle: video.title, video });
 };
 
 export const getEdit = async (req, res) => {
     const { id } = req.params;
+    const {
+        user: { _id },
+    } = req.session;
     const video = await Video.findById(id);
     if (!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." });
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     return res.render("edit", { pageTitle: `Edit ${video.title}`, video });
 };
 
 export const postEdit = async (req, res) => {
     const { id } = req.params;
+    const {
+        user: { _id },
+    } = req.session;
     const video = await Video.exists({ _id: id });
     if (!video) {
         return res.render("404", { pageTitle: "Video not found." });
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     const { title, description, hashtags } = req.body;
     await Video.findByIdAndUpdate(id, {
@@ -52,18 +64,21 @@ export const postUpload = async (req, res) => {
         file: { path: fileUrl },
         body: { title, description, hashtags },
         session: {
-            user: { _id: owner },
+            user: { _id },
         },
     } = req;
 
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
             fileUrl,
             hashtags: Video.formatHashtags(hashtags),
-            owner,
+            owner: _id,
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
         return res.redirect("/");
     } catch (error) {
         return res.status(400).render("upload", {
@@ -75,6 +90,16 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const { id } = req.params;
+    const {
+        user: { _id },
+    } = req.session;
+    const video = await Video.exists({ _id: id });
+    if (!video) {
+        return res.render("404", { pageTitle: "Video not found." });
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 };
